新規に追加したBusinessクラスとそのテストのエラーを解消。 テストクラスで必要になるIsSerializableを仮で用意。

+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/ShotokuIdentifier.java
@@ -6,6 +6,9 @@
 package jp.co.ndensan.reams.db.dbz.business.core;
 
 import java.io.Serializable;
+import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
+import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import lombok.Value;
 
 /**
@@ -16,7 +19,7 @@ public class ShotokuIdentifier implements Serializable {
 
     private final FlexibleYear 所得年度;
     private final ShikibetsuCode 識別コード;
-    private final int 履歴番号;
+    private final Decimal 履歴番号;
 
     /**
      * コンストラクタです。
@@ -26,8 +29,8 @@ public class ShotokuIdentifier implements Serializable {
      * @param 履歴番号 履歴番号
      */
     public ShotokuIdentifier(FlexibleYear 所得年度,
-ShikibetsuCode 識別コード,
-int 履歴番号) {
+            ShikibetsuCode 識別コード,
+            Decimal 履歴番号) {
         this.所得年度 = 所得年度;
         this.識別コード = 識別コード;
         this.履歴番号 = 履歴番号;
